% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Holdfasts-descr.R
\docType{data}
\name{Holdfasts}
\alias{Holdfasts}
\title{Kelp holdfast data}
\format{
Dataframe of dimension 80 by 51. Holdfast density (per sq m),  depth (m) and holdfast volume (ml) are in the first three columns, and species counts in the remaining 48 columns.
}
\usage{
data(Holdfasts)
}
\description{
Counts of 48 taxon inhabiting kelp holdfasts at 80 sites.
}
\examples{
data(Holdfasts)
}
\references{
Millar, Anderson & Zunan (2005) Fitting nonlinear environmental gradients to community data: A general distance based approach. Ecology. 86: 2245-2251)
}
\keyword{datasets}
