% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplineCAP.R
\name{LOO.prednError}
\alias{LOO.prednError}
\title{Calculates the leave-one-out prediction sums of squares of a multivariate linear regression}
\usage{
LOO.prednError(f.object)
}
\arguments{
\item{f.object}{Model formula}
}
\value{
A list containing residual sums of squares and leave-one-out prediction sums of squares
}
\description{
Calculates the leave-one-out prediction sums of squares of a multivariate linear regression
}
\examples{
data(PitTraps)
Locn=PitTraps$Locn
distance=PitTraps$distance
Counts=PitTraps[,-c(1,2)]

B=bSpline(distance,degree=2,df=5)   #Spline basis for distance
First need to calculate PCOs, as these are the Y variables
D=as.matrix(vegdist(Counts>0)) #Bray-Curtis is default in vegdist
Y.pco=cmdscale(D,k=nrow(D)-1,eig=T)$points #The PCOs
LOO.prednError(Y.pco~B+Locn)

}
