% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PitTraps-descr.R
\docType{data}
\name{PitTraps}
\alias{PitTraps}
\title{Pitfall trap data}
\format{
Dataframe of dimension 120 by 189. Location ("B", "R", "C" or "L") and distance (m) from habitat boundary are in the first two columns, and species counts in the remaining 187 columns.
}
\usage{
data(PitTraps)
}
\description{
Counts of 187 terrestial invertebrate species caught in pitfall traps spanning a habitat boundary.
}
\examples{
data(PitTraps)
}
\references{
Millar, Anderson & Zunan (2005) Fitting nonlinear environmental gradients to community data: A general distance based approach. Ecology. 86: 2245-2251)
}
\keyword{datasets}
